feat(backend): add initial best_combination service using a greedy set cover algorithm

+++ backend/src/common/models/streaming_offer_schema.rs
@@ -2,16 +2,16 @@ use mongodb::bson::oid;
 use serde::{Deserialize, Serialize};
 use validator::Validate;
 
-#[derive(Serialize, Deserialize, Debug, Validate)]
+#[derive(Clone, Serialize, Deserialize, Debug, Validate)]
 pub struct StreamingOffer {
     #[serde(rename = "_id")]
     pub id: oid::ObjectId,
 
     #[validate(range(min = 1))]
-    pub game_id: u8,
+    pub game_id: u32,
 
     #[validate(range(min = 1))]
-    pub streaming_package_id: u8,
+    pub streaming_package_id: u32,
 
     #[validate(range(min = 0, max = 1))]
     pub live: u8,
